Reorganized project to use react-router package
